pyproject.toml
[build-system]
requires = ["flit_core >=3.7.1,<4"]
build-backend = "flit_core.buildapi"


[project]
name = "bigdict"
authors = [
    {name = "Zepu Zhang", email = "zepu.zhang@gmail.com"},
]
dependencies = [
    "lmdb",
]
requires-python = ">=3.8"
readme = "README.rst"
license = {file = "LICENSE"}
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = ['version', 'description']


[project.urls]
Source = "https://github.com/zpz/bigdict"


[project.optional-dependencies]
rocksdb = [
    # "lbry-rocksdb",
    "rocksdb",
]
doc = [
    "sphinx",
    "numpydoc",
    "pydata-sphinx-theme",
]
test = [
    "mypy",
    "ruff",
    "pytest-asyncio",
]


[tool.ruff]
target-version = "py38"
select = ["E", "F", "S", "I001"]  # isort
ignore = ["E501", "S101", "S102", "S103", "S104", "S108", "S301", "S311", "S608"]
exclude = ["tests/data010"]


[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-sv --log-cli-level info -p no:cacheprovider --tb=short --ignore=tests/data010 --durations 3"
asyncio_mode = "auto"


[tool.coverage.report]
fail_under = 90
show_missing = true
skip_empty = true


[tool.coverage.run]
source = ["bigdict"]
data_file = "/tmp/.coverage"
